+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 ### Added
 - Added nbest translation, exposed as `--nbest-size`. Nbest translation means to not only output the most probable translation according to a model, but the top n most probable hypotheses. If `--nbest-size > 1` and the option `--output-type` is not explicitly specified, the output type will be changed to one JSON list of nbest translations per line. `--nbest-size` can never be larger than `--beam-size`.
 
+### Changed
+- Changed `sockeye.rerank` CLI to be compatible with nbest translation JSON output format.
+
 ## [1.18.57]
 ### Added
 - Added `sockeye.score` CLI for quickly scoring existing translations ([documentation](tutorials/scoring.md)).